refactor(OptionsConfigurationComponent): rename from ConfigurationComponent
